remove method Explorer.openSettingPane
